Moved some classes around, final commit before release
