+++ honeypot-core/src/main/java/org/reprogle/honeypot/utils/HoneypotLogger.java
@@ -8,7 +8,6 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
 import org.reprogle.honeypot.Honeypot;
-import org.reprogle.honeypot.HoneypotConfigManager;
 
 public class HoneypotLogger {
